CON-622: Move `panoramaOfJustifications` to constructor of `MessageJPast`.
