+++ casper/src/main/scala/io/casperlabs/casper/util/DagOperations.scala
@@ -510,109 +510,7 @@ object DagOperations {
     message.justifications.toList
       .map(_.latestBlockHash)
       .traverse(dag.lookupUnsafe(_))
-      .flatMap(panoramaOfMessageFromJustifications[F](dag, _, erasObservedBehavior, stop))
-
-  def panoramaOfMessageFromJustifications[F[_]: MonadThrowable](
-      dag: DagLookup[F],
-      justifications: List[Message],
-      erasObservedBehavior: LocalDagView[Message],
-      stop: Message
-  ): F[MessageJPast[Message]] = {
-    type PerEraValidatorMessages =
-      Map[ByteString, Map[Validator, ObservedValidatorBehavior[Message]]]
-    import ObservedValidatorBehavior._
-    implicit val blockTopoOrdering: Ordering[Message] =
-      DagOperations.blockTopoOrderingDesc
-
-    val stream = toposortJDagDesc(dag, justifications)
-      .takeUntil(_ == stop)
-      .filter(!_.isGenesisLike) // Not interested in Genesis block
-
-    val keyBlockHashes = erasObservedBehavior.keyBlockHashes
-
-    val initStates: PerEraValidatorMessages =
-      keyBlockHashes.toList
-        .map(
-          kbh =>
-            kbh -> erasObservedBehavior
-              .validatorsInEra(kbh)
-              .filterNot(_.isEmpty()) // Not interested in Genesis validator.
-              .map(_ -> ObservedValidatorBehavior.Empty)
-              .toMap
-        )
-        .toMap
-
-    // Previously seen message from the validator.
-    val prevMessageSeen: Map[ByteString, Map[Validator, Message]] = Map.empty
-    val validatorStatusMatches
-        : ByteString => Validator => ObservedValidatorBehavior[Message] => Boolean =
-      era => validator => seenBehavior => erasObservedBehavior.data(era)(validator) == seenBehavior
-
-    // If we've seen, in the j-past-cone of the message, the same statuses
-    // as ones we have collected locally then we're done.
-    // This is correct b/c a node has to download all dependencies first,
-    // before validating the block.
-    val isDone: PerEraValidatorMessages => Boolean = state =>
-      state.forall {
-        case (era, eraLMS) =>
-          eraLMS.forall {
-            case (validator, seenStatus) =>
-              validatorStatusMatches(era)(validator)(seenStatus)
-          }
-      }
-
-    stream
-      .foldWhileLeft((initStates, prevMessageSeen)) {
-        case ((latestMessages, prevMessages), message) =>
-          if (isDone(latestMessages)) {
-            Right((latestMessages, prevMessages))
-          } else {
-            val msgCreator = message.validatorId
-            val msgEra     = message.eraId
-            val eraMsgs    = latestMessages(msgEra)
-            Left(eraMsgs(msgCreator) match {
-              case Honest(prevMsg) =>
-                if (prevMessages(msgEra)(msgCreator).validatorPrevMessageHash == message.messageHash)
-                  (
-                    latestMessages,
-                    prevMessages
-                      .updated(msgEra, prevMessages(msgEra).updated(msgCreator, message))
-                  )
-                else {
-                  // Validator equivocated in this era
-                  (
-                    latestMessages.updated(
-                      msgEra,
-                      latestMessages(msgEra).updated(msgCreator, Equivocated(prevMsg, message))
-                    ),
-                    prevMessages
-                  )
-                }
-              case Empty =>
-                (
-                  latestMessages
-                    .updated(msgEra, latestMessages(msgEra).updated(msgCreator, Honest(message))),
-                  prevMessages
-                    .updated(
-                      msgEra,
-                      prevMessages.getOrElse(msgEra, Map.empty).updated(msgCreator, message)
-                    )
-                )
-              case Equivocated(_, _) => (latestMessages, prevMessages)
-            })
-          }
-      }
-      .map {
-        case (a, _) =>
-          EraObservedBehavior.messageJPast(
-            a.mapValues(_.mapValues {
-              case Empty               => Set.empty
-              case Honest(m)           => Set(m)
-              case Equivocated(m1, m2) => Set(m1, m2)
-            })
-          )
-      }
-  }
+      .flatMap(EraObservedBehavior.messageJPast[F](dag, _, erasObservedBehavior, stop))
 
   /**
     * Returns latest messages per era.
